Add history API endpoint
The endpoint returns a JSON array of palette definitions, in descending order of most recently used. The contents are updated at build time, adding a palette which is not already present. The length is capped at half the number of enabled palettes.

This endpoint is designed to be consumed by the skyline generation script, which should only use the 0th element. This will ensure that palettes will not reappear on the bot's timeline until a significant number of other palettes have been used.